requirements.txt
BCEmbedding==0.1.5
BCEmbedding==0.1.5
faiss_cpu==1.11.0
fastapi==0.115.12
langchain==0.3.25
numpy==1.26.4
openai==1.79.0
pydantic==2.11.4
Requests==2.32.3
streamlit==1.45.1